Sauvegarde du 06/02/2025, Creation entity user + review de l entity, model, dao, repository, interface repository, use case. Creation user 1 avec room pour afficher photo de profile sur DetailLayout.

+++ app/src/main/java/com/nedrysystems/joiefull/data/entity/ProductEntity.kt
@@ -17,7 +17,7 @@ data class ProductEntity(
     @PrimaryKey
     val id: Int,
     val favorite: Boolean = false,
-    val rate: Double? = null
+    val rate: Int? = null
 
 )
